removed the hasVoted booleans and instead use the userVote number property to make checks. postVote() combines what earlier was two separate methods postVote() & removeVote()
when a user clicks the icon, it triggers postVote(). if the uservote is 0 that means a vote has not been made yet. So vote. If a userVote is the type of vote clicked, remove that vote. If the userVote is not the type of vote clicked, then remove the vote and add a new one with the new type