when editing, I use a loader gif as boolean specific for each field, bio, username, email, that turn off after updateUser() returns successfully.
ngOnInit also sets the user service user to the currently loggedinuser. necessary?

after update is success, the response.body is set to three different users. I really wish it was just one. Will need to refactor?

userprofile gets a user, but if i'm on my profile, the user is me, the loggedinuser. If i'm just browsing someone elses profile, the user is the service user.